fix creating projects and transcript
issue using fetch to create an HTTP resource with json payload in body is a different request from creating a resource with a form data that has a media file, eg you cannot stringify the media file. Needs further looking into the fetch config that I've commented out to get it working for now.